Decouple formatting from toot text
This allows even more robust truncation, as the text of the toot is
no longer pre-quoted. The output function is now in charge of adding the
quotation marks, and therefore it is also in charge of handling what to
do at the end of a line when the text is too long.